// employee-happiness-app-frontend/src/app/employee-happiness/interfaces/employee-happiness.interface.ts
export interface WorkEnvironment {
    envRadio: number;
	envButtonToggle: number;
	envSlider1: number;
	envSlider2: number 
  }
  
  export interface BenefitsAndCompensation{
    benefitRadio1: number;
	benefitButtonToggle: number;
	benefitRadio2: number;
	benefitSlider: number;
  }

  export interface CareerDevelopment{
    careerRadio1: number;
    careerButtonToggle: number;
    careerRadio2: number;
  }